v1.05.040 debuging

+++ src/main/resources/initShopContent.sql
@@ -203,7 +203,7 @@ VALUES (1, CONCAT(@UploadsUser,"user.jpeg")),
 -- add roles
 DELETE FROM roles WHERE id >= 0;
 INSERT INTO roles(id, name)
-VALUES (1, "ROLE_USER"), (2, "ROLE_ADMIN");
+VALUES (1, "USER"), (2, "ADMIN");
 
 -- add users
 DELETE FROM users WHERE id >= 0;
@@ -211,7 +211,7 @@ INSERT INTO users(name, surname, email, password, date_of_birth, balance, image_
 VALUES ("Admin", "Admin", "admin", "admin", "1980-01-01", 9999999, 1);
 
 -- add users_roles
-DELETE FROM users_roles;
+DELETE FROM users_roles WHERE user_id >= 0;
 INSERT INTO users_roles(user_id, role_id)
 VALUES (1, 2);
